Add Weekdays Two more components created: Weekday and Day. Weekday controls getting the all the days in the current month and displaying them in a column structure
