+++ src/calendarData/calendarData.js
@@ -1,6 +1,6 @@
 export const calendarData = {
   months: ["Januray", "February", "March", "April", "May", "June", "July", "August", "September", "October", "November", "December"],
-  weekday: ["Sunday", "Monay", "Tuesday", "Wednesday", "Thursday", "Friday", "Saturday", "Sunday"],
+  weekday: ["Sunday", "Monday", "Tuesday", "Wednesday", "Thursday", "Friday", "Saturday", "Sunday"],
   mockEvents: [
     {
       start: '8:30am',
