+++ src/components/Day/Day.js
@@ -1,0 +1,9 @@
+import React, { Component } from 'react';
+
+export const Day = ({ day }) => {
+  return (
+    <div className="Day">
+      <h3>{ day }</h3>
+    </div>
+  )
+}
